Add main class, documentation and corrected mapping

+++ src/main/java/ru/yandex/clickhouse/jdbcbridge/db/clickhouse/ClickHouseConverter.java
@@ -36,9 +36,9 @@ public class ClickHouseConverter {
 
     static {
         Map<Integer, MappingInstruction> map = new HashMap<>();
-        map.put(INTEGER, new MappingInstruction<>(Int32, ResultSet::getInt, (i, s) -> s.writeInt32(i)));
+        map.put(TINYINT, new MappingInstruction<>(Int8, ResultSet::getInt, (i, s) -> s.writeUInt8(i)));
         map.put(SMALLINT, new MappingInstruction<>(Int16, ResultSet::getInt, (i, s) -> s.writeInt16(i)));
-        map.put(TINYINT, new MappingInstruction<>(UInt8, ResultSet::getInt, (i, s) -> s.writeUInt8(i)));
+        map.put(INTEGER, new MappingInstruction<>(Int32, ResultSet::getInt, (i, s) -> s.writeInt32(i)));
         map.put(BIGINT, new MappingInstruction<>(Int64, ResultSet::getLong, (i, s) -> s.writeInt64(i)));
 
         map.put(FLOAT, new MappingInstruction<>(Float32, ResultSet::getFloat, (i, s) -> s.writeFloat32(i)));
